editarPaciente y cambios en controladores
