+++ MedicApp/src/main/java/com/ProyectoFinal/MedicApp/controller/PacienteControlador.java
@@ -12,6 +12,7 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.http.HttpSession;
@@ -31,6 +32,7 @@ import org.springframework.web.bind.annotation.RequestParam;
  */
 
 @Controller
+@PreAuthorize("hasAnyRole('ROLE_PACIENTE')")
 @RequestMapping("/pac")
 public class PacienteControlador {
     
@@ -40,54 +42,14 @@ public class PacienteControlador {
     @Autowired
     PacienteService pacienteService;
 
-    @GetMapping("/form_pac")
-    public String form_pac(ModelMap model) {
-        return "formulario_paciente.html";
-    }
-    
-    @GetMapping("/listar_pac")
-    public String listar_pac(ModelMap model) {
-        List<Paciente> pacientes = pacienteService.listar();
-        model.addAttribute("pacientes", pacientes);
-        return "listar_pac.html";
-    }
-
-    @PostMapping("/registroPaciente")
-    public String registroPaciente(@RequestParam String nombre, @RequestParam String apellido,
-            @RequestParam String correo, @RequestParam String telefono, @RequestParam String nacimiento,
-            @RequestParam String password, @RequestParam String password2, @RequestParam String direccion,
-            @RequestParam String sexo) {
-       
-        try {
-            SimpleDateFormat formato = new SimpleDateFormat("yyyy-MM-dd");
-            Date fechaNacimiento = formato.parse(nacimiento);
-            
-            pacienteService.crearPaciente(nombre, apellido, correo, telefono, password, password2, direccion, fechaNacimiento, sexo);
-            System.out.println("Ingreso de paciente exitoso");
-            return "redirect:/inicio";
-            
-        } catch (MiExcepcion me) {
-            System.out.println("Ingreso de paciente FALLIDO!\n" + me.getMessage());
-            
-            return "formulario_paciente.html";
-            
-        } catch (ParseException ex) {
-            System.out.println(ex.getMessage());
-            ex.printStackTrace();
-            return "formulario_paciente.html";
-        }
-    }
-    
-    @PreAuthorize("hasAnyRole('ROLE_PACIENTE')")
     @GetMapping("/perfil")
     public String perfil(ModelMap modelo, HttpSession session) {
         Paciente paciente = (Paciente) session.getAttribute("pacienteSession");
-        modelo.put("pacienteSession", paciente);
+        modelo.put("paciente", paciente);
         
-        return "formulario_paciente.html";
+        return "editar_paciente.html";
     }
     
-    @PreAuthorize("hasAnyRole('ROLE_PACIENTE')")
     @PostMapping("/perfil/{id}")
     public String modificarPerfil(@PathVariable String id, @RequestParam String nombre, @RequestParam String apellido,
             @RequestParam String correo, @RequestParam String telefono, @RequestParam String nacimiento,
@@ -102,17 +64,25 @@ public class PacienteControlador {
             String password2, String direccion, Date fechaNacimiento, String sexo*/
             pacienteService.modificarPaciente(id, nombre, apellido, correo, telefono, password, password2, direccion, fechaNacimiento, sexo);
             session.setAttribute("pacienteSession", pacienteService.getOne(id));
-            return "redirect:/pac/perfil";
+            return "redirect:/inicio";
             
         } catch (MiExcepcion me) {
             System.out.println("Ingreso de paciente FALLIDO!\n" + me.getMessage());
             
-            return "formulario_paciente.html";
+            return "ediar_paciente.html";
             
         } catch (ParseException ex) {
             System.out.println(ex.getMessage());
             ex.printStackTrace();
-            return "formulario_paciente.html";
+            return "editar_paciente.html";
         }
+    }
+    
+    @GetMapping("/baja/{id}")
+    public String bajaPaciente(@PathVariable String id) {
+        
+        pacienteService.darDeBaja(id);
+        
+        return "redirect:/";
     }
 }
