+++ MedicApp/src/main/resources/templates/login.html
@@ -59,7 +59,7 @@
           
           <div class="divCrearUsuario">
             <p>Si todavía no tenés usuario</p>
-            <a href="/pac/form_pac" id="linkCrear" class="link"
+            <a href="/form_pac" id="linkCrear" class="link"
               ><button class="btn btn-primary">Hacé click acá</button></a
             >
           </div>
